MLLM backend: put utility functions in separate annif.lexical.util module
