+++ annif/lexical/mllm.py
@@ -6,14 +6,14 @@ import joblib
 from statistics import mean
 from enum import IntEnum
 import numpy as np
-from rdflib import URIRef
 from rdflib.namespace import SKOS
-from scipy.sparse import lil_matrix, csc_matrix
 from sklearn.feature_extraction.text import CountVectorizer
 from sklearn.ensemble import BaggingClassifier
 from sklearn.tree import DecisionTreeClassifier
 import annif.util
 from annif.lexical.tokenset import TokenSet, TokenSetIndex
+from annif.lexical.util import get_subject_labels
+from annif.lexical.util import make_relation_matrix, make_collection_matrix
 
 
 Term = collections.namedtuple('Term', 'subject_id label is_pref')
@@ -32,13 +32,6 @@ Feature = IntEnum(
     'first_occ last_occ spread doc_length ' +
     'broader narrower related collection',
     start=0)
-
-
-def get_subject_labels(graph, uri, properties, language):
-    for prop in properties:
-        for label in graph.objects(URIRef(uri), prop):
-            if label.language == language:
-                yield str(label)
 
 
 class MLLMModel:
@@ -128,43 +121,14 @@ class MLLMModel:
 
         return (terms, subject_ids)
 
-    def _make_relation_matrix(self, graph, vocab, property):
-        n_subj = len(vocab.subjects)
-        matrix = lil_matrix((n_subj, n_subj), dtype=np.bool)
-
-        for subj, obj in graph.subject_objects(property):
-            subj_id = vocab.subjects.by_uri(str(subj), warnings=False)
-            obj_id = vocab.subjects.by_uri(str(obj), warnings=False)
-            if subj_id is not None and obj_id is not None:
-                matrix[subj_id, obj_id] = True
-
-        return csc_matrix(matrix)
-
-    def _make_collection_matrix(self, graph, vocab):
-        # make an index with all collection members
-        c_members = collections.defaultdict(list)
-        for coll, member in graph.subject_objects(SKOS.member):
-            member_id = vocab.subjects.by_uri(str(member), warnings=False)
-            if member_id is not None:
-                c_members[str(coll)].append(member_id)
-
-        c_matrix = lil_matrix((len(c_members), len(vocab.subjects)),
-                              dtype=np.bool)
-
-        # populate the matrix for collection -> subject_id
-        for c_id, members in enumerate(c_members.values()):
-            c_matrix[c_id, members] = True
-
-        return csc_matrix(c_matrix)
-
     def _prepare_relations(self, graph, vocab):
-        self._broader_matrix = self._make_relation_matrix(
+        self._broader_matrix = make_relation_matrix(
             graph, vocab, SKOS.broader)
-        self._narrower_matrix = self._make_relation_matrix(
+        self._narrower_matrix = make_relation_matrix(
             graph, vocab, SKOS.narrower)
-        self._related_matrix = self._make_relation_matrix(
+        self._related_matrix = make_relation_matrix(
             graph, vocab, SKOS.related)
-        self._collection_matrix = self._make_collection_matrix(graph, vocab)
+        self._collection_matrix = make_collection_matrix(graph, vocab)
 
     def _prepare_train_index(self, vocab, analyzer, params):
         graph = vocab.as_graph()
